+++ templates/end_users/end_user.html
@@ -151,7 +151,7 @@
             <br><br>
 
             <div class="lite-tabs__container">
-                <a href="{% url 'end_users:end_user_detail' end_user_advisory.id 'case-notes' %}" class="lite-tabs__tab {% if type == 'case-notes' %}lite-tabs__tab--selected{% endif %}" id="case-notes-tab">
+                <a href="{% url 'end_users:end_user_detail' end_user_advisory.id 'case-notes' %}" class="lite-tabs__tab {% if type == 'case-notes' %}lite-tabs__tab--selected{% endif %}" id="link-case-notes">
                     Notes
                     {% if case_note_notifications %}
                         <div class="lite-notification-bubble">
@@ -159,7 +159,7 @@
                         </div>
                     {% endif %}
                 </a>
-                <a href="{% url 'end_users:end_user_detail' end_user_advisory.id 'ecju-queries' %}" class="lite-tabs__tab {% if type == 'ecju-queries' %}lite-tabs__tab--selected{% endif %}" id="ecju-queries-tab">
+                <a href="{% url 'end_users:end_user_detail' end_user_advisory.id 'ecju-queries' %}" class="lite-tabs__tab {% if type == 'ecju-queries' %}lite-tabs__tab--selected{% endif %}" id="link-ecju-queries">
                     ECJU Queries
                     {% if ecju_query_notifications %}
                         <div class="lite-notification-bubble">
