+++ templates/goods/good.html
@@ -191,7 +191,7 @@
 		    {% if good.query_id %}
 				<br><br>
 		        <div class="lite-tabs__container">
-		            <a href="{% url 'goods:good_detail' good.id 'case-notes' %}" class="lite-tabs__tab {% if type == 'case-notes' %}lite-tabs__tab--selected{% endif %}" id="case-notes-tab">
+		            <a href="{% url 'goods:good_detail' good.id 'case-notes' %}" class="lite-tabs__tab {% if type == 'case-notes' %}lite-tabs__tab--selected{% endif %}" id="link-case-notes">
 					    Notes
 		                {% if case_note_notifications %}
 		                    <div class="lite-notification-bubble">
@@ -199,7 +199,7 @@
 		                    </div>
 		                {% endif %}
 				    </a>
-		            <a href="{% url 'goods:good_detail' good.id 'ecju-queries' %}" class="lite-tabs__tab {% if type == 'ecju-queries' %}lite-tabs__tab--selected{% endif %}" id="ecju-queries-tab">
+		            <a href="{% url 'goods:good_detail' good.id 'ecju-queries' %}" class="lite-tabs__tab {% if type == 'ecju-queries' %}lite-tabs__tab--selected{% endif %}" id="link-ecju-queries">
 		                ECJU Queries
 		                {% if ecju_query_notifications %}
 		                    <div class="lite-notification-bubble">
